simplify quarkus sidecar code
